requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
selenium==4.27.1
webdriver-manager==4.0.2
python-dotenv==1.0.1
pydantic==2.10.4
loguru==0.7.3
httpx==0.28.1
